Time: 26 ms (73.14%) | Memory: 30 MB (12.54%) - LeetSync

+++ 792-binary-search/binary-search.cpp
@@ -7,7 +7,7 @@ public:
         }
 
 
-    long long mid  = si+((ei-si)%2);
+    int mid  = si+((ei-si)%2);
 
     if( nums[mid] == target ) {
         return mid;
